requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
alembic==1.8.0
    # via flask-migrate
async-timeout==4.0.2
    # via redis
certifi==2022.6.15
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via flask
colorama==0.4.5
    # via click
deprecated==1.2.13
    # via redis
dnspython==2.2.1
    # via -r requirements.in
flask==2.1.2
    # via
    #   -r requirements.in
    #   flask-migrate
    #   flask-redis
    #   flask-sqlalchemy
flask-migrate==3.1.0
    # via -r requirements.in
flask-redis==0.4.0
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.in
    #   flask-migrate
greenlet==1.1.2
    # via sqlalchemy
idna==3.3
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
mistune==2.0.2
    # via -r requirements.in
mysqlclient==2.1.1
    # via -r requirements.in
packaging==21.3
    # via redis
pycryptodome==3.15.0
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
python-dotenv==0.20.0
    # via -r requirements.in
redis==4.3.3
    # via flask-redis
requests==2.28.0
    # via -r requirements.in
sqlalchemy==1.4.39
    # via
    #   alembic
    #   flask-sqlalchemy
urllib3==1.26.9
    # via requests
waitress==2.1.2
    # via -r requirements.in
werkzeug==2.1.2
    # via flask
wrapt==1.14.1
    # via deprecated
